requirements.txt
Flask==3.1.0
tensorflow==2.18.0
pillow==11.0.0
gunicorn==20.1.0
numpy==2.0.2
flask_cors==5.0.0
